# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API De Tuna Resort",
        "description": "Dokumentasi API untuk Sistem Informasi & Pemesanan De Tuna Resort",
        "contact": {
            "email": "admin@detuna.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server API De Tuna (Lokal - php artisan serve)"
        },
        {
            "url": "http://detuna-api.test",
            "description": "Server API De Tuna (Lokal - Laragon)"
        }
    ],
    "paths": {
        "/api/v1/admin/bookings": {
            "get": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Get semua data booking (Admin)",
                "operationId": "3638e4bacbb3a5a5b3656adfbcd9ac83",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status booking (pending, confirmed, canceled, completed)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar semua booking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Booking"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Token tidak valid)"
                    },
                    "403": {
                        "description": "Forbidden (Tidak memiliki peran Admin)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookings/{id}/status": {
            "put": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Update status booking (Admin)",
                "operationId": "d16120958fdeaecbdd9e2091dac26182",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Status baru untuk booking",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_status"
                                ],
                                "properties": {
                                    "booking_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "canceled",
                                            "completed"
                                        ],
                                        "example": "confirmed"
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "failed"
                                        ],
                                        "example": "paid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status booking berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal (status tidak valid)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/messages": {
            "get": {
                "tags": [
                    "Admin - Messages"
                ],
                "summary": "Get semua pesan (Admin)",
                "operationId": "cb8f8d4bb8fbb2c6dd4265a6e33a3df8",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "read",
                                "unread"
                            ]
                        },
                        "example": "unread"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pesan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Message"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/messages/{id}": {
            "get": {
                "tags": [
                    "Admin - Messages"
                ],
                "summary": "Get detail pesan (Admin)",
                "operationId": "b49efc5353b5a64a322f02afe38d800e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pesan"
                    },
                    "404": {
                        "description": "Pesan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Messages"
                ],
                "summary": "Hapus pesan",
                "operationId": "038e5f0cb6c565c1205dae6740510e03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesan berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/messages/{id}/read": {
            "put": {
                "tags": [
                    "Admin - Messages"
                ],
                "summary": "Tandai pesan sebagai 'sudah dibaca'",
                "operationId": "7e4c7aae1363fe7ea47b2092b634eda4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesan ditandai dibaca"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "title": "Booking",
                "description": "Model Pemesanan",
                "properties": {
                    "booking_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "booking_code": {
                        "type": "string",
                        "example": "DT-20251021-XYZ"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-20"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-22"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 5000000
                    },
                    "booking_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "canceled",
                            "completed"
                        ],
                        "example": "pending"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "failed"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Message": {
                "title": "Message",
                "description": "Model Pesan Kontak",
                "properties": {
                    "message_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sender_name": {
                        "type": "string",
                        "example": "Budi"
                    },
                    "sender_email": {
                        "type": "string",
                        "format": "email",
                        "example": "budi@example.com"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Pertanyaan"
                    },
                    "message_body": {
                        "type": "string",
                        "example": "Apakah tersedia..."
                    },
                    "is_read": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "sanctum": {
            "type": "http",
            "description": "Masukkan Bearer Token (Sanctum)",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}